feat: semantic analysis base
This has a lot of changes. A lot of refactor things to ease
development of pop, the semantic analisis through scopes
has been initialized. With that I know that I'll reach my goal of
finish this project. Token now has its own file,
SymbolTable and Scopes are established, log library has improved
and a lot more of little changes.