feat: compilation of simple mathematic expressions
This is the first step to achieve this project.
Next thing is comparisons to be able to implement if statements.
This code adds a module of generation where is going to be any
implementation of the backend to any type of assembly.
I want to compile it into WASM and ARM. But first x86_64 with nasm.