feat: variables are available now
It was easier than I thought, but it's tricky, they are only global
variables and I need to implement the local ones.
I don't now yet how I'm going to reach that, but I will.
Also I need to refactor the code because it's getting a little difficult
to work with.